Use system fonts stack
Custom fonts may look fancy and make a website more personal, but
ultimately we should strive for consistency and just use the system
fonts that the user configured instead of blindly overwriting them.

This also makes the website load faster and avoids connecting to Google
servers, which is always a big plus.

For now this uses the font-family definition from modernfontstacks.

refs: https://modernfontstacks.com/
refs: https://blog.jim-nielsen.com/2020/system-fonts-on-the-web/
refs: https://systemfontstack.com/
refs: https://fonts.google.com/knowledge/glossary/system_font_web_safe_font