Add an option to inline all CSS
This is still a bit buggy [0], so this is off by default. We should turn
this on by default once it works without problems.

The idea is that we then (and with favicon inlined) only have one single
file to serve for each page.

Still we need to investigate client compatibility, because in
8217ac0c4edeb128c52de53013ae6c992294a7be we already had to revert
inlining of favicons by default.

[0] https://github.com/getzola/zola/issues/605#issuecomment-1867642699