Export accent color
Some browsers (e.g. Chromium on Android) make use of the theme color to
highlight the tab bar in a different color.

For now we hardcode the same accent color as "--primary-color" in the
CSS files. In the future we should allow the consumer to configure the
accent color in the configuration file and get the value from there.

Unfortunately this is not possible at the moment, because zola is
missing support for CSS templating, but an implementation is already in
the pipeline [0], so with some luck we might soon be able to unify the
accent color retrieval.

[0] https://github.com/getzola/zola/pull/2242