Block LLM scrapers
A recent thread in lobsters [0] made me aware of this list, and thanks
to the dark visitors site [1] we can block all known LLM scrapers.

For the public good, let's add this to the Zola robots.txt template [2]
which is picked up from the templates directory [3], keep in mind though
that this can simply be circumvented by setting a different User-Agent
or by ignoring robots.txt, so it is not really effective.
Blocking at the nginx level is not really more effective too though.

I have never given OpenAI permission to reuse my CC-BY-SA-4.0 content in
their glorified garbage neural network and to burp it out to users
losing the license in the process, so this may help in preventing that
from happening in the first place.

[0] https://lobste.rs/s/ybowdq/great_gpt_firewall
[1] https://darkvisitors.com/
[2] https://github.com/getzola/zola/blob/master/components/templates/src/builtins/robots.txt
[3] https://www.getzola.org/documentation/templates/robots/